project-1/Project - 3/requirements.txt
Flask==3.0.2
google-generativeai==0.8.3